Update fr.csv : Second correction
Correction of the first translation after building the app and understanding better what each things means with the context.
I've also added "saved", "Discount After Tax" and "New Entry" translation to the file, i'll add them to the other languages later. and add the translation code in an other branch dedicated to improve the translated text in the app to not see english bits remaining anymore.